Adding the information viewer
Added information viewer to the client, where he can ask for the information then see it.
also fixed the known bug from previous commit with the bad dice being used when getting pieces out.
also, made the end position of the pips look better

+++ Backgammon V2/Assets/Scripts/User.cs
@@ -348,7 +348,7 @@ namespace BackGammonUser
 
         public void GetLatestUserInformation()
         {
-            AddDataToSend(MessageType.InformationContainer.ToString(), MessageType.RequestData);
+            AddDataToSend(((int)MessageType.InformationContainer).ToString(), MessageType.RequestData);
         }
 
         public void StartNewGame()
